Update README.md
This is the GitHub repository for Stat 360, spring 2022 at SFU. The lecture notes and many of the exercises are based on the text Advanced R by Hadley Wickham. These materials are licensed under the same license as the source material.

Creative Commons License
This work is licensed under a Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License.